backend/requirements.txt
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-filter>=23.5
django-cors-headers>=4.3.0
django-extensions>=3.2.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
pytz>=2024.1
pillow>=10.2.0
boto3>=1.34.0
django-storages>=1.14.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
pyotp>=2.9.0
qrcode>=7.4.0
gunicorn>=21.2.0
